Update MiddlemanServer.java
added Scanner object for input of port number to open a room. Created a do/While loop to ensure a free port is able to be accessed and made a method to check if a port is free.

also set up location for close room method as I believe the server should be able to close a room in the case of it not being in use.